My_Cynocular_Nginx_app/requirements.txt
Flask-Cors
Flask-SQLAlchemy
flask_jwt_extended
python-decouple
Flask-Migrate
Werkzeug